rename jump_cache.rs -> jump_analysis.rs & move analysis into it
