+++ rust/src/types/code_reader.rs
@@ -4,12 +4,12 @@ use std::{
     ops::Deref,
 };
 
-use crate::types::{code_byte_type, u256, CodeByteType, FailStatus, JumpAnalysis, Opcode};
+use crate::types::{u256, CodeByteType, FailStatus, JumpAnalysis, Opcode};
 
 #[derive(Debug)]
 pub struct CodeReader<'a> {
     code: &'a [u8],
-    code_byte_types: JumpAnalysis,
+    jump_analysis: JumpAnalysis,
     pc: usize,
 }
 
@@ -31,7 +31,7 @@ impl<'a> CodeReader<'a> {
     pub fn new(code: &'a [u8], code_hash: Option<u256>, pc: usize) -> Self {
         Self {
             code,
-            code_byte_types: JumpAnalysis::new(code_hash, || compute_code_byte_types(code)),
+            jump_analysis: JumpAnalysis::new(code, code_hash),
             pc,
         }
     }
@@ -39,7 +39,7 @@ impl<'a> CodeReader<'a> {
     pub fn get(&self) -> Result<Opcode, GetOpcodeError> {
         if self.pc >= self.code.len() {
             Err(GetOpcodeError::OutOfRange)
-        } else if self.code_byte_types[self.pc] == CodeByteType::DataOrInvalid {
+        } else if self.jump_analysis[self.pc] == CodeByteType::DataOrInvalid {
             Err(GetOpcodeError::Invalid)
         } else {
             let op = self.code[self.pc];
@@ -58,9 +58,7 @@ impl<'a> CodeReader<'a> {
 
     pub fn try_jump(&mut self, dest: u256) -> Result<(), FailStatus> {
         let dest = u64::try_from(dest).map_err(|_| FailStatus::BadJumpDestination)? as usize;
-        if dest >= self.code_byte_types.len()
-            || self.code_byte_types[dest] != CodeByteType::JumpDest
-        {
+        if dest >= self.jump_analysis.len() || self.jump_analysis[dest] != CodeByteType::JumpDest {
             return Err(FailStatus::BadJumpDestination);
         }
         self.pc = dest;
@@ -84,119 +82,12 @@ impl<'a> CodeReader<'a> {
     }
 }
 
-fn compute_code_byte_types(code: &[u8]) -> Vec<CodeByteType> {
-    let mut code_byte_types = vec![CodeByteType::DataOrInvalid; code.len()];
-
-    let mut pc = 0;
-    while pc < code.len() {
-        let (code_byte_type, inc) = code_byte_type(code[pc]);
-        code_byte_types[pc] = code_byte_type;
-        pc += inc;
-    }
-
-    code_byte_types
-}
-
 #[cfg(test)]
 mod tests {
     use crate::types::{
-        code_reader::{compute_code_byte_types, CodeReader, GetOpcodeError},
-        u256, CodeByteType, FailStatus, Opcode,
+        code_reader::{CodeReader, GetOpcodeError},
+        u256, FailStatus, Opcode,
     };
-
-    #[test]
-    fn compute_code_byte_types_single_byte() {
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::Add as u8]),
-            [CodeByteType::Opcode]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::Push2 as u8]),
-            [CodeByteType::Opcode]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::JumpDest as u8]),
-            [CodeByteType::JumpDest]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[0xc0]),
-            [CodeByteType::DataOrInvalid]
-        );
-    }
-
-    #[test]
-    fn compute_byte_types_jumpdest() {
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::JumpDest as u8, Opcode::Add as u8]),
-            [CodeByteType::JumpDest, CodeByteType::Opcode]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::JumpDest as u8, 0xc0]),
-            [CodeByteType::JumpDest, CodeByteType::DataOrInvalid]
-        );
-    }
-
-    #[test]
-    fn compute_code_byte_types_push_with_data() {
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::Push1 as u8, Opcode::Add as u8, Opcode::Add as u8]),
-            [
-                CodeByteType::Opcode,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::Opcode,
-            ]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[Opcode::Push1 as u8, Opcode::Add as u8, 0xc0]),
-            [
-                CodeByteType::Opcode,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::DataOrInvalid,
-            ]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[
-                Opcode::Push1 as u8,
-                Opcode::Add as u8,
-                0xc0,
-                Opcode::Add as u8
-            ]),
-            [
-                CodeByteType::Opcode,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::Opcode,
-            ]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[
-                Opcode::Push2 as u8,
-                Opcode::Add as u8,
-                Opcode::Add as u8,
-                Opcode::Add as u8,
-            ]),
-            [
-                CodeByteType::Opcode,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::Opcode,
-            ]
-        );
-        assert_eq!(
-            *compute_code_byte_types(&[
-                Opcode::Push2 as u8,
-                Opcode::Add as u8,
-                Opcode::Add as u8,
-                0xc0
-            ]),
-            [
-                CodeByteType::Opcode,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::DataOrInvalid,
-                CodeByteType::DataOrInvalid,
-            ]
-        );
-    }
 
     #[test]
     fn code_reader_internals() {
