+++ rust/src/types/mod.rs
@@ -4,7 +4,7 @@ mod cache;
 mod code_reader;
 mod execution_context;
 pub mod hash_cache;
-mod jump_cache;
+mod jump_analysis;
 mod memory;
 mod mock_execution_message;
 mod opcode;
@@ -17,7 +17,7 @@ pub use amount::u256;
 pub use cache::Cache;
 pub use code_reader::{CodeReader, GetOpcodeError};
 pub use execution_context::*;
-pub use jump_cache::JumpAnalysis;
+pub use jump_analysis::JumpAnalysis;
 pub use memory::Memory;
 pub use mock_execution_message::MockExecutionMessage;
 pub use opcode::*;
